==> Retención de cartera de clientes utilizando ensambles de machine learning/notebooks/4. Pipeline/2. Armado de pagos.feather.ipynb <==
import pandas as pd
pd.set_option("display.max_columns", 500)
pd.set_option("display.max_rows", 500)

table = "PAGOS.tsv"
location = "../../../data/raw"
sep = '\t'
encoding = 'latin1'
decimal = ','

## Checking columns to remove from sample

for chunk in pd.read_csv(f"{location}/{table}",
                         sep=sep,
                         encoding=encoding,
                         decimal=decimal,
                         chunksize=5000000):
    df = chunk
    break

sorted(df.columns)

usable_cols = ['COD_CIA',
 'COD_COBRO',
 'COD_SECC',
 'COD_SITUACION',
 'FECHA_EQUIPO',
 'FECHA_VTO',
 'FEC_PAGO',
 'MONTO_PAGO',
 'NUM_END',
 'NUM_SECU_POL']

"], [".join(usable_cols)

## Creating pagos.feahter

chunks = pd.read_csv(f"{location}/{table}",
                     sep=sep,
                     encoding=encoding,
                     decimal=decimal,
                     chunksize=1000000,
                     iterator=True,
                     usecols = usable_cols)
df = pd.concat(chunks)

fechas = ["FECHA_EQUIPO","FECHA_VTO","FEC_PAGO"]

df[fechas].head(5)

Transformamos FECHA_EQUIPO a formato año-mes

def lookup(s):
    """
    This is an extremely fast approach to datetime parsing.
    For large data, the same dates are often repeated. Rather than
    re-parse these, we store all unique dates, parse them, and
    use a lookup to convert all dates.
    """
    dates = {date:pd.to_datetime(date).strftime('%Y-%m') for date in s.unique()}
    return s.map(dates)

df["FECHA_EQUIPO"] = lookup(df["FECHA_EQUIPO"])

df[fechas].head(5)

df.dtypes

df.to_feather("../../../data/interim/pagos.feather")
