==> Retención de cartera de clientes utilizando ensambles de machine learning/notebooks/4. Pipeline/3. Armado de clientes.feather.ipynb <==
import pandas as pd
pd.set_option("display.max_columns", 500)
pd.set_option("display.max_rows", 500)

table = "TB_CIF.tsv"
location = "../../../data/raw"
sep = '\t'
encoding = 'latin1'
decimal = ','

## Checking columns to remove from sample

for chunk in pd.read_csv(f"{location}/{table}",
                         sep=sep,
                         encoding=encoding,
                         decimal=decimal,
                         chunksize=5000000):
    df = chunk
    break

sorted(df.columns)

usable_cols = [
 'ANO_DE_NACIMIENTO',
 'CLIENTE',
 'DATECO_AUTONOMO',
 'DATECO_RAZON_SOCIAL_EMPLEADOR',
 'DATECO_TRABAJA',
 'DATMAR_COMPLETO_ESTUDIOS',
 'DATMAR_NIVEL_EDUCACION',
 'ESTADO_CIVIL',
 'FECHA_NACIMIENTO',
 'ID',
 'ORIGEN',
 'POSICION_IVA',
 'SEXO',
 'TIPO_PERSONA']


all_  =['ANO_DE_NACIMIENTO',
 'CLIENTE',
 'DATECO_AUTONOMO',
 'DATECO_INGRESO_FAMILIAR',
 'DATECO_INGRESO_PERSONAL',
 'DATECO_RAZON_SOCIAL_EMPLEADOR',
 'DATECO_TIPO_ACTIVIDAD',
 'DATECO_TRABAJA',
 'DATMAR_CANTIDAD_HIJOS',
 'DATMAR_COMPLETO_ESTUDIOS',
 'DATMAR_NIVEL_EDUCACION',
 'DOMICILIO_CODIGO_POSTAL',
 'DOMICILIO_LATITUD',
 'DOMICILIO_LONGITUD',
 'ESTADO_CIVIL',
 'FECHA_DEFUNCION',
 'FECHA_NACIMIENTO',
 'HABILITADO',
 'ID',
 'NACIONALIDAD',
 'ORIGEN',
 'POSICION_IB',
 'POSICION_IVA',
 'SEXO',
 'TIPO_PERSONA',
 'VALIDO']

non_usable = [x for x in all_ if x not in usable_cols]

"], [".join(usable_cols)

"], [".join(non_usable)

## Creating clientes.feahter

chunks = pd.read_csv(f"{location}/{table}",
                     sep=sep,
                     encoding=encoding,
                     decimal=decimal,
                     chunksize=1000000,
                     iterator=True,
                     usecols = usable_cols)
df = pd.concat(chunks)

df.to_feather("../../../data/interim/clientes.feather")
