==> Retención de cartera de clientes utilizando ensambles de machine learning/notebooks/4. Pipeline/5. Armado de poliza_ID.feather.ipynb <==
import pandas as pd

table = "VIGABT_POLIZAS.tsv"
location = "../../../data/raw"
sep = '\t'
encoding = 'latin1'
decimal = ','

df_polizas = pd.read_csv(f"{location}/{table}",
                         sep=sep,
                         encoding=encoding,
                         decimal=decimal,
                         usecols = ["NUM_SECU_POL", "CIF_ID"])

df_polizas.head()

df_polizas.drop_duplicates(inplace=True)

df_polizas.reset_index(drop=True).to_feather("../../../data/interim/poliza_x_cliente.feather")
