==> Retención de cartera de clientes utilizando ensambles de machine learning/notebooks/4. Pipeline/11. Creacion de siniestros por CIF_ID.ipynb <==
import pandas as pd

df_sini = pd.read_feather("../../../data/interim/siniestros.feather")

df_ids = pd.read_feather("../../../data/interim/poliza_x_cliente.feather")

Verificación de cruce

len(set(df_ids["NUM_SECU_POL"]).intersection(set(df_sini["NUM_SECU_POL"]))) / len(df_ids)

len(set(df_sini["NUM_SECU_POL"]).intersection(set(df_ids["NUM_SECU_POL"]))) / len(df_sini)

### Agregamos CIF_ID a los Siniestros

df_sini_id = pd.merge(df_sini, df_ids, on=['NUM_SECU_POL'], how='left')

df_sini_id.drop_duplicates(subset=["NUM_SECU_POL", "FEC_DENU_SINI"], keep='last', inplace=True)

### Agrupamos siniestros por periodo

data = {fecha:i + 1 for i,fecha in enumerate(sorted(df_sini_id["FEC_DENU_SINI"].unique()))}

data

df_sini_id["periodo_denu_sini"] = df_sini_id["FEC_DENU_SINI"].map(data)

df_sini_id["periodo_liquidacion_sini"] = df_sini_id["FECHA_LIQUIDACION"].map(data)

df_sini_id["periodo_rechazo_sini"] = df_sini_id["FECHA_RECHAZO"].map(data)

to_pivot = df_sini_id[["CIF_ID","NUM_SECU_POL","TIPO_SINIESTRO","periodo_denu_sini","periodo_liquidacion_sini","periodo_rechazo_sini"]]

### Pivots por CIF_ID

pd.set_option("display.max_columns", 500)

df_sini = to_pivot.pivot_table(index='CIF_ID',columns=['periodo_denu_sini'], values=['NUM_SECU_POL','periodo_liquidacion_sini','periodo_rechazo_sini'],aggfunc='count', fill_value=0)

df_sini.head()

df_tipo = to_pivot.pivot_table(index='CIF_ID',columns=['periodo_denu_sini','TIPO_SINIESTRO'],aggfunc='size', fill_value=0)

df_tipo.head()

Reagrupamos los DF

df_sini = pd.DataFrame(df_sini.to_records())
df_tipo = pd.DataFrame(df_tipo.to_records())

df_sini.head()

df_tipo.head()

len(df_tipo) == len(df_sini)

** Cambio de nombre de columnas **

df_tipo.columns

df_tipo = df_tipo.rename(columns=lambda x: x.replace("(","").replace(")","").replace(", ","_TIPO_SINI_").replace("'",""))

df_tipo.head()

df_sini.columns

df_sini = df_sini.rename(columns=lambda x: x.replace("(","").replace(")","").replace(", ","_").replace("'","").replace("NUM_SECU_POL","periodo_sini"))

df_sini.columns

### Ahora unimos los dataframes

df = pd.merge(df_sini, df_tipo, on=['CIF_ID'], how='outer')

len(df_sini.columns) + len(df_tipo.columns)

len(df.columns)

df.to_feather("../../../data/processed/siniestros_x_cif_id.feather")
