==> Retención de cartera de clientes utilizando ensambles de machine learning/notebooks/4. Pipeline/12. Merge de Bajas, Pagos, Interacciones, Clientes, Siniestros.ipynb <==
import pandas as pd
pd.set_option("display.max_columns", 500)
pd.set_option("display.max_rows", 500)

## Importamos los archivos

df_bajas = pd.read_feather('../../../data/processed/periodos_baja_x_cif_id.feather')

df_pagos = pd.read_feather('../../../data/processed/pagos_x_cif_id.feather')

df_interacciones = pd.read_feather('../../../data/processed/interacciones_tipo_periodo_x_cif_id.feather')

df_clientes = pd.read_feather('../../../data/interim/clientes.feather')

df_clientes = df_clientes.rename(columns={"ID":"CIF_ID"})

df_siniestros = pd.read_feather('../../../data/processed/siniestros_x_cif_id.feather')

## Verificamos que todos se puedan combinar con la tabla Bajas

def check_merge(df1,df2,field):
    return len(set(df1[field].unique()).intersection(set(df2[field].unique()))) / len(df1[field].unique())

check_merge(df_bajas, df_pagos, "CIF_ID")

check_merge(df_bajas, df_clientes, "CIF_ID")

*Renombrando columnas de interacciones y descartando periodos que no vamos a analizar*

df_interacciones.columns

df_interacciones = df_interacciones.rename(columns=lambda x: x.replace("(","").replace(")","").replace(", ","_TIPOINT_").replace("'","")) 

to_drop = [col for col in df_interacciones.drop('CIF_ID', axis=1).columns if int(col.split('_')[0])>18]

to_drop

df_interacciones.drop(to_drop, axis=1, inplace=True)

check_merge(df_bajas, df_interacciones, "CIF_ID")

check_merge(df_bajas, df_siniestros, "CIF_ID")

## Merge entre tablas

len(df_bajas)

len(df_interacciones)

len(df_pagos)

len(df_clientes)

len(df_siniestros)

df_bajas_int =  pd.merge(df_bajas, df_interacciones, on=['CIF_ID'], how='left')

df_bajas_int.head()

df_bajas_int_sini =  pd.merge(df_bajas_int, df_siniestros, on=['CIF_ID'], how='left')

df_bajas_int_sini = df_bajas_int_sini.fillna(0)

df_bajas_int_sini.head()

from functools import reduce
df_merged = reduce(lambda  left,right: pd.merge(left, right, on=['CIF_ID'], how='inner'), [df_bajas_int_sini, df_pagos, df_clientes])

df_merged.shape

### Guardamos la tabla completa

df_merged.to_feather('../../../data/processed/merged_data_18_periods.feather')

--- 

import pandas as pd
df_merged = pd.read_feather('../../../data/processed/merged_data_18_periods.feather')

"], [".join(set([''.join([s for s in col if not s.isdigit()]) for col in df_merged.columns]))
